+++ src/card/cards/combat_constraints.rs
@@ -24,3 +24,4 @@ pub const TELEPORT: CardDefinitionId = CardDefinitionId(1823);
 pub const RESET: CardDefinitionId = CardDefinitionId(1824);
 pub const LIVING_ARTIFACT: CardDefinitionId = CardDefinitionId(1825);
 pub const INSTILL_ENERGY: CardDefinitionId = CardDefinitionId(1826);
+pub const HOWLING_MINE: CardDefinitionId = CardDefinitionId(1827);