+++ src/game/tests.rs
@@ -570,6 +570,7 @@ mod fetch_tapped;
 mod filtered_player_prevention;
 mod guardian_beast;
 mod held_tapped;
+mod howling_mine;
 mod identity_and_mana;
 mod instill_energy;
 mod isd_dgm_deck_cards;